Migrated to Pyannote for speaker diarization

+++ transcribe.nf
@@ -38,20 +38,18 @@ process to_wav {
 
 process diarization {
     memory '5GB'
+    label 'with_gpu'
     
     input:
     path audio
 
     output: 
-    path 'show.seg', emit: show_seg, optional: true
-    path 'show.uem.seg', emit: show_uem_seg 
+    path 'show.rttm', emit: show_rttm    
 
     script:        
         """
-        find_speech_segments.py $audio show.uem.seg
-        if [ -s show.uem.seg ]; then
-          diarization.sh $audio show.uem.seg
-        fi
+        diarize.py $audio show.0.rttm
+        postprocess_rttm.py show.0.rttm show.rttm
         """
 }
 
@@ -59,34 +57,59 @@ process prepare_initial_data_dir {
     memory '1GB'
 
     input:
-      path show_seg
+      path show_rttm
       path audio
 
     output:
       path 'init_datadir', emit: init_datadir, optional: true
 
     shell:
-    '''
-    . !{projectDir}/bin/prepare_process.sh
-    
-    if [ -s !{show_seg} ]; then
-      mkdir init_datadir
-      echo audio audio A > init_datadir/reco2file_and_channel
-      cat !{show_seg} | cut -f "3,4,8" -d " " | \
-      while read LINE ; do \
-        len_in_frames=`echo $LINE | cut -f 2 -d " "`; \
-        start=`echo $LINE | cut -f 1,2 -d " " | perl -ne '@t=split(); $start=$t[0]/100.0; printf("%09.3f", $start);'`; \
-        end=`echo $LINE   | cut -f 1,2 -d " " | perl -ne '@t=split(); $start=$t[0]/100.0; $len=$t[1]/100.0; $end=$start+$len; printf("%09.3f", $end);'`; \
-        sp_id=`echo $LINE | cut -f 3 -d " "`; \
-        if  [ ${len_in_frames} -gt 30 ]; then \
-          echo audio-${sp_id}---${start}-${end} audio $start $end; \
-        fi
-      done > init_datadir/segments
-      cat init_datadir/segments | perl -npe 's/\\s+.*//; s/((.*)---.*)/\\1 \\2/' > init_datadir/utt2spk
-      utils/utt2spk_to_spk2utt.pl init_datadir/utt2spk > init_datadir/spk2utt
-      echo "audio !{audio}" > init_datadir/wav.scp
-      
-    fi
+    '''    
+    if [ $(wc -l < !{show_rttm}) -gt 1 ]; then
+        . !{projectDir}/bin/prepare_process.sh
+        mkdir init_datadir
+        echo audio audio A > init_datadir/reco2file_and_channel
+        
+        while read -r line; do
+            # Skip lines that do not start with "SPEAKER"
+            if [[ ! $line =~ ^SPEAKER ]]; then
+                continue
+            fi
+
+            # Read fields from RTTM line
+            IFS=' ' read -r -a fields <<< "$line"
+            file_id=audio
+            channel=${fields[2]}
+            start_time=${fields[3]}
+            duration=${fields[4]}
+            speaker_id=${fields[7]}
+
+            # Skip segments shorter than 0.3 seconds
+            if (( $(echo "$duration < 0.3" | bc -l) )); then
+                continue
+            fi
+
+            # Calculate end time
+            end_time=$(echo "$start_time + $duration" | bc -l)
+
+            # Format start and end times with zero-padding
+            start_time_padded=$(LC_NUMERIC="C" printf "%08.3f" $start_time)
+            end_time_padded=$(LC_NUMERIC="C" printf "%08.3f" $end_time)
+
+            # Generate unique utterance ID
+            utt_id="${file_id}_${speaker_id}_${start_time_padded}-${end_time_padded}"
+
+            # Append to segments and utt2spk files
+            echo "$utt_id $file_id $start_time $end_time" >> init_datadir/segments
+            echo "$utt_id $speaker_id" >> init_datadir/utt2spk
+
+        done < !{show_rttm}
+          
+        echo "audio !{audio}" > init_datadir/wav.scp  
+        utils/utt2spk_to_spk2utt.pl init_datadir/utt2spk > init_datadir/spk2utt
+        echo "audio !{audio}" > init_datadir/wav.scp
+    fi      
+
     '''
 }
 
@@ -209,8 +232,9 @@ process decode_whisper {
       #whisper-transcribe.py --beam 1 --is_multilingual --language et !{params.rootdir}/models/whisper-large-et-v3-ct2 !{audio} datadir_25sec/segments | tee trans.hyp
       #whisper-transcribe.py --beam 1 --is_multilingual --language et !{params.rootdir}/models/whisper-medium-et-v4-ct2 !{audio} datadir_25sec/segments | cut -f 2- -d " " | perl -npe \'s/(\\S)([,.!?:])/\\1 \\2/g\' | paste <(cut -f 1 -d " " datadir_25sec/segments) - | tee trans.hyp
       
-      #whisper-ctranslate2 --language et --model_directory !{params.rootdir}/models/whisper-medium-et-v5-ct2 !{audio} --beam_size 2
-      whisper-ctranslate2 --device cuda --language et --model_directory  !{params.rootdir}/models/whisper-large-et-orth-v2-ct2 !{audio} --beam_size  1 --vad_filter True
+      whisper-ctranslate2 --vad_filter True --language et --model_directory !{params.rootdir}/models/whisper-medium-et-v5-ct2 !{audio} --beam_size 2
+      #whisper-ctranslate2 --device cuda --language et --model_directory  !{params.rootdir}/models/whisper-large-et-orth-v2-ct2 !{audio} --beam_size  1 --vad_filter True
+      #whisper-ctranslate2 --device cuda --language et --model_directory  !{params.rootdir}/models/whisper-large-et-orth-20240510-ct2 !{audio} --beam_size  3 --vad_filter True
       
       '''
 }
@@ -228,15 +252,16 @@ process postprocess_whisper {
       
     shell:
       '''
-      . !{projectDir}/bin/prepare_process.sh
-      mkdir datadir_whisper     
-      echo audio audio A > datadir_whisper/reco2file_and_channel
-      echo "audio audio.wav" > datadir_whisper/wav.scp      
-      tail -n +2 !{audio_tsv} | awk '{printf("%04d", NR); print(" audio " $1/1000 " " $2/1000);}' > datadir_whisper/segments
-      tail -n +2 !{audio_tsv} | awk '{printf("%04d", NR); $1=""; $2=""; print }' > datadir_whisper/text
-      awk '{print($1, $1)}' datadir_whisper/segments > datadir_whisper/utt2spk
-      awk '{print($1, $1)}' datadir_whisper/segments > datadir_whisper/spk2utt
-      
+      if [ $(wc -l < !{audio_tsv}) -gt 1 ]; then
+          . !{projectDir}/bin/prepare_process.sh
+          mkdir datadir_whisper     
+          echo audio audio A > datadir_whisper/reco2file_and_channel
+          echo "audio audio.wav" > datadir_whisper/wav.scp      
+          tail -n +2 !{audio_tsv} | awk '{printf("%04d", NR); print(" audio " $1/1000 " " $2/1000);}' > datadir_whisper/segments
+          tail -n +2 !{audio_tsv} | awk '{printf("%04d", NR); $1=""; $2=""; print }' > datadir_whisper/text
+          awk '{print($1, $1)}' datadir_whisper/segments > datadir_whisper/utt2spk
+          awk '{print($1, $1)}' datadir_whisper/segments > datadir_whisper/spk2utt
+      fi      
       '''
 }
 
@@ -289,7 +314,7 @@ process align {
       #cat !{datadir_whisper}/text | perl -npe 's/ [[:punct:]]/ /g;' >  datadir_whisper_hires/text      
       steps/make_mfcc.sh --nj 1 \
           --mfcc-config !{params.rootdir}/kaldi-data/!{params.acoustic_model}/conf/mfcc.conf \
-          datadir_whisper_hires || exit 1
+          datadir_whisper_hires || echo "Warning: steps/make_mfcc.sh resulted in error"
       steps/compute_cmvn_stats.sh datadir_whisper_hires || exit 1
       utils/fix_data_dir.sh datadir_whisper_hires
 
@@ -308,7 +333,7 @@ process align {
       utils/prepare_lang.sh --phone-symbol-table !{params.rootdir}/kaldi-data/!{params.acoustic_model}/phones.txt --sil-prob 0.01 dict '<unk>' dict/tmp lang
       cp -r !{params.rootdir}/kaldi-data/!{params.acoustic_model}/conf .      
 
-      steps/nnet3/align.sh --retry-beam 100 --use-gpu true --nj 1 --online-ivector-dir ivectors \
+      steps/nnet3/align.sh --retry-beam 300 --use-gpu true --nj 1 --online-ivector-dir ivectors \
         --scale-opts '--transition-scale=1.0 --self-loop-scale=1.0 --acoustic-scale=1.0' \
         datadir_whisper_hires lang !{params.rootdir}/kaldi-data/!{params.acoustic_model} ali
         
@@ -322,18 +347,20 @@ process align {
 }
 
 
+
 process output {
     memory '500MB'
     
-    publishDir "${out_dir}${basename}", mode: 'copy', overwrite: true
+    publishDir "${out_dir}/${basename}", mode: 'copy', overwrite: true
 
     input:
       val basename
       path datadir
-      path show_uem_seg
+      path show_rttm
       path sid_result
       path alignments_json
       path alignments_ctm
+      path empty_result_txt
 
     output:
       path "result.json"
@@ -344,11 +371,11 @@ process output {
 
     shell:
       '''
-      resegment_json.py --new-turn-sil-length 1.0 --speaker-names !{sid_result} --pms-seg !{show_uem_seg} !{datadir} !{alignments_json} > punctuated.json
+      resegment_json.py --new-turn-sil-length 1.0 --speaker-names !{sid_result} --rttm !{show_rttm} !{datadir} !{alignments_json} > punctuated.json
       
       normalize_json.py words2numbers.py punctuated.json > result.json
 
-      json2trs.py --fid trs punctuated.json > result.trs
+      json2trs.py --fid trs result.json > result.trs
       json2srt.py result.json > result.srt
       cat !{alignments_ctm} | perl -npe 's/[.,!?;:]$//' >  result.ctm
       json2text.py < result.json > result.txt
@@ -356,22 +383,21 @@ process output {
       '''
 }
 
+
 process empty_output {
 
     publishDir "${out_dir}/${basename}", mode: 'copy', overwrite: true
 
     input:
-      val a from datadir.ifEmpty{ 'EMPTY' }
-      val b from init_datadir.ifEmpty{ 'EMPTY' }
-    when:
-      a == 'EMPTY' || b == 'EMPTY' 
+      val basename
 
     output:
-      file "result.json" into empty_result_json
-      file "result.srt" into empty_result_srt
-      file "result.trs" into empty_result_trs
-      file "result.ctm" into empty_result_ctm
-      file "result.txt" into empty_result_txt
+      path "result.json", emit: empty_result_json
+      path "result.srt", emit: empty_result_srt
+      path "result.trs", emit: empty_result_trs
+      path "result.ctm", emit: empty_result_ctm
+      path "result.txt", emit: empty_result_txt
+      
 
     script:
       json = file("assets/empty.json")
@@ -388,12 +414,13 @@ workflow {
   files = params.in_file_list != "" ? Channel.fromPath(params.in_file_list).splitText() : Channel.fromPath(params.in)
   
   to_wav(files)
+  empty_output(to_wav.out.basename)
   diarization(to_wav.out.audio)
-  prepare_initial_data_dir(diarization.out.show_seg, to_wav.out.audio)
+  prepare_initial_data_dir(diarization.out.show_rttm, to_wav.out.audio)
   language_id(prepare_initial_data_dir.out.init_datadir, to_wav.out.audio)
   speaker_id(language_id.out.datadir, to_wav.out.audio)
   decode_whisper(to_wav.out.audio)
   postprocess_whisper(decode_whisper.out.audio_tsv)
   align(postprocess_whisper.out.datadir_whisper, to_wav.out.audio)
-  output(to_wav.out.basename, language_id.out.datadir, diarization.out.show_uem_seg, speaker_id.out.sid_result, align.out.alignments_json, align.out.alignments_ctm)
+  output(to_wav.out.basename, language_id.out.datadir, diarization.out.show_rttm, speaker_id.out.sid_result, align.out.alignments_json, align.out.alignments_ctm,empty_output.out.empty_result_txt)
 }
